+++ packages/@tinacms/cli/src/next/commands/dev-command/server/index.ts
@@ -1,13 +1,12 @@
-import bodyParser from 'body-parser'
-import path from 'path'
-import cors from 'cors'
 import { createServer as createViteServer } from 'vite'
 import type { Plugin } from 'vite'
 import { resolve as gqlResolve, Database } from '@tinacms/graphql'
 import { ConfigManager } from '../../../config-manager'
-import { parseMediaFolder, createMediaRouter } from './media'
-import { transform as esbuildTransform } from 'esbuild'
 import { createConfig } from '../../../vite'
+import {
+  devServerEndPointsPlugin,
+  transformTsxPlugin,
+} from '../../../vite/plugins'
 
 export const createDevServer = async (
   configManager: ConfigManager,
@@ -17,83 +16,8 @@ export const createDevServer = async (
   noWatch: boolean
 ) => {
   const plugins: Plugin[] = [
-    {
-      name: 'transform-tsx',
-      async transform(code, id) {
-        // FIXME: this seems to only happen in the monorepo
-        // so it's not used in the vite build fro now
-        //
-        // Vite isn't picking up the right transform for tsx
-        // files, we could opt-out of this if the file isn't
-        // .tsx but seems to work ok for now.
-        // TODO: other loaders needed (eg svg)?
-        if (id.startsWith(configManager.rootPath)) {
-          const extName = path.extname(id)
-          if (extName.startsWith('.tsx') || extName.startsWith('.ts')) {
-            const result = await esbuildTransform(code, { loader: 'tsx' })
-            return {
-              code: result.code,
-            }
-          }
-        }
-      },
-    },
-    {
-      name: 'graphql-endpoints',
-      configureServer(server) {
-        server.middlewares.use(cors())
-        server.middlewares.use(bodyParser.json())
-        server.middlewares.use(async (req, res, next: Function) => {
-          const mediaPaths = configManager.config.media?.tina
-          const mediaRouter = createMediaRouter({
-            rootPath: configManager.rootPath,
-            apiURL,
-            publicFolder: parseMediaFolder(mediaPaths?.publicFolder || ''),
-            mediaRoot: parseMediaFolder(mediaPaths?.mediaRoot || ''),
-          })
-          if (req.url.startsWith('/media/upload')) {
-            await mediaRouter.handlePost(req, res)
-            return
-          }
-          if (req.url.startsWith('/media')) {
-            if (req.method === 'DELETE') {
-              await mediaRouter.handleDelete(req, res)
-              return
-            }
-          }
-          if (req.url.startsWith('/media/list')) {
-            await mediaRouter.handleList(req, res)
-            return
-          }
-          if (req.url === '/altair') {
-            res.end(
-              JSON.stringify({
-                status:
-                  'The GraphQL playground has moved to <your-dev-url>/index.html#/graphql',
-              })
-            )
-            return
-          }
-          if (req.url === '/graphql') {
-            // @ts-ignore FIXME: req type doesn't match
-            const { query, variables } = req.body
-            const result = await gqlResolve({
-              config: {
-                useRelativeMedia: true,
-              },
-              database,
-              query,
-              variables,
-              verbose: false,
-            })
-            res.end(JSON.stringify(result))
-            return
-          }
-
-          next()
-        })
-      },
-    },
+    transformTsxPlugin({ configManager }),
+    devServerEndPointsPlugin({ apiURL, configManager, database }),
   ]
   return createViteServer(
     await createConfig({
